rename hooks filename
