+++ packages/nextjs/hooks/scaffold-eth/index.ts
@@ -5,8 +5,8 @@ export * from "./useNativeCurrencyPrice";
 export * from "./useNetworkColor";
 export * from "./useOutsideClick";
 export * from "./useScaffoldContract";
-export * from "./useScaffoldContractRead";
-export * from "./useScaffoldContractWrite";
+export * from "./useScaffoldReadContract";
+export * from "./useScaffoldWriteContract";
 export * from "./useScaffoldEventSubscriber";
 export * from "./useScaffoldEventHistory";
 export * from "./useTransactor";
