+++ packages/nextjs/hooks/scaffold-eth/useScaffoldReadContract.ts
