+++ packages/nextjs/hooks/scaffold-eth/useScaffoldWriteContract.ts
